add support for "inheriting" transformations among rules/attributes
i. e. if A is transformed by f, A has B as a dependant
and B is transformed by g, then B will be transformed to f(g(B))
and B's dependants likewise

in particular this allows the following (in config-file-speech) to work:
A f
   B
B
   C
C will also be transformed by f, which would not have been the case before